Fix typo in Invasion Engine description
Starting to wonder if i should work towards seperating out the "sample game" parts from core engine parts. Or if they will forever be intertwined and if people want to actually make their own game out of it they shoudl just change whats here. But start with the sample as a base? 

Signed-off-by: Untrustedlife <bountybayonlineroblox@gmail.com>